feat(machinery): extract component as machinery page for the usage of management and establishment
