+++ src/Components/MachineryManagement/RemarksModal.tsx
@@ -23,11 +23,11 @@ export default function RemarksTable(props: {
     isOpen: boolean;
     onClose: () => void;
     info: IMachinery;
-    readOnly: boolean;
+    editable: boolean;
 }) {
-    const { isOpen, onClose, info, readOnly } = props;
+    const { isOpen, onClose, info, editable } = props;
     const { inspectionNo, siteId, remarks } = info;
-    const [editable, setEditable] = React.useState<boolean>(false);
+    const [editMode, setEditMode] = React.useState<boolean>(false);
     const [originText, setOriginText] = React.useState<string>(
         remarks.text || ''
     );
@@ -79,7 +79,7 @@ export default function RemarksTable(props: {
         <Modal
             isOpen={isOpen}
             onClose={() => {
-                setEditable(false);
+                setEditMode(false);
                 onClose();
             }}
             size={'xl'}
@@ -98,7 +98,7 @@ export default function RemarksTable(props: {
                             >
                                 備註
                             </Text>
-                            {!readOnly && (
+                            {!editable && (
                                 <IconButton
                                     size={'xs'}
                                     h={'20px'}
@@ -107,7 +107,7 @@ export default function RemarksTable(props: {
                                     aria-label="edit remarks"
                                     icon={<EditIcon />}
                                     onClick={() => {
-                                        setEditable(true);
+                                        setEditMode(true);
                                     }}
                                 />
                             )}
@@ -133,7 +133,7 @@ export default function RemarksTable(props: {
                                 opacity: 1,
                                 bg: '#FFFFFF',
                             }}
-                            disabled={!editable}
+                            disabled={!editMode}
                         ></Textarea>
                         <Text
                             variant={'w700s16'}
@@ -153,7 +153,7 @@ export default function RemarksTable(props: {
                                 listType="picture"
                                 fileList={remarksPhotos}
                                 autoUpload={false}
-                                removable={editable}
+                                removable={editMode}
                                 draggable
                                 multiple
                                 action="#"
@@ -174,7 +174,7 @@ export default function RemarksTable(props: {
                                     </Box>
                                 )}
                             >
-                                {editable ? (
+                                {editMode ? (
                                     <Box
                                         style={{
                                             margin: '0px',
@@ -198,7 +198,7 @@ export default function RemarksTable(props: {
                                 )}
                             </Uploader>
                         </Flex>
-                        {editable ? (
+                        {editMode ? (
                             <Flex justify={'space-between'} width={'100%'}>
                                 <Button
                                     variant={'buttonGrayOutline'}
@@ -206,7 +206,7 @@ export default function RemarksTable(props: {
                                     height={'36px'}
                                     mr={3}
                                     onClick={() => {
-                                        setEditable(false);
+                                        setEditMode(false);
                                         setRemarksText(originText);
                                         setRemarksPhotos(originPhotos);
                                     }}
@@ -229,7 +229,7 @@ export default function RemarksTable(props: {
                                                 ),
                                             },
                                         }).then(() => {
-                                            setEditable(false);
+                                            setEditMode(false);
                                             setRemarksText(
                                                 remarksRef.current?.value || ''
                                             );
